feat: implement GeoJSON worker for asynchronous parsing and processing
